Fix build warnings, improve timezone picker UX, and add CLAUDE.md
- SecondTimezonePickerView: replace List with LazyVStack in ScrollView, add remove button above list, improve row hit targets
- PlantShape: explicit CGFloat cast for cos/sin to silence type warnings
- TableInsertView: use simplified NSTextTableBlock border API
- Minor single-line fixes in AppDelegate, RichTextCoordinator, RichTextEditor
- Add CLAUDE.md with architecture overview and build commands

+++ StickyBloom/PlantAnimation/PlantShape.swift
@@ -172,8 +172,8 @@ struct PlantShape: Shape {
             for i in 0..<plant.petalCount {
                 let angle = (Double(i) / Double(plant.petalCount)) * .pi * 2
                 let petalCenter = CGPoint(
-                    x: center.x + cos(angle) * radius * 0.7,
-                    y: center.y + sin(angle) * radius * 0.7
+                    x: center.x + CGFloat(cos(angle)) * radius * 0.7,
+                    y: center.y + CGFloat(sin(angle)) * radius * 0.7
                 )
                 path.addEllipse(in: CGRect(
                     x: petalCenter.x - petalRadius,
